Fixes small configuration glitches
Fixes self-tracing, removes the storage type registry, and replaces
"zipkin.store.type" with "zipkin.storage.type" which closer relates to
the env variable "STORAGE_TYPE"
